+++ R/install.R
@@ -11,6 +11,18 @@ NULL
 #' Download specified version of EnergyPlus for your platform from GitHub and
 #' install it.
 #'
+#' @param ver The EnergyPlus version number, e.g., \code{8.7}; the special value
+#' \code{"latest"} means the latest version (fetched from GitHub releases).
+#'
+#' @param force Whether to install EnergyPlus even if it has already been
+#' installed.
+#'
+#' @param dir Where to save EnergyPlus installer file. Default is current
+#' working directory for `download_eplus()` and `tempdir()` for
+#' `install_eplus()`.
+#'
+#' @details
+#'
 #' `download_eplus()` downloads specified version of EnergyPlus from
 #' [EnergyPlus Github Repository](https://github.com/NREL/EnergyPlus).
 #'
@@ -21,22 +33,13 @@ NULL
 #'
 #' Note that the installation process requires administrative privileges
 #' during the installation and you have to run R with administrator (or with
-#' sudo if you are on Linux) to make it work.
-#'
-#' @param ver The EnergyPlus version number, e.g., \code{8.7}; the special value
-#' \code{"latest"} means the latest version (fetched from GitHub releases).
-#'
-#' @param dir Where to save EnergyPlus installer file. Default is current
-#' working directory. Only applicable to `download_eplus()`. For
-#' `install_eplus()`, it will be saved into `tempdir()`.
-#'
-#' @param force Whether to install EnergyPlus even if it has already been
-#' installed.
+#' sudo if you are on Linux) to make it work if you are not in interactive mode.
 #'
 #' @name install_eplus
 #' @export
+#'
 # install_eplus {{{
-install_eplus <- function (ver = "latest", force = FALSE) {
+install_eplus <- function (ver = "latest", force = FALSE, dir = tempdir()) {
     # check if the same version has been installed already
     if (ver == "latest") ver <- eplus_latest_release()
     ver_exists <- is_avail_eplus(ver)
@@ -52,15 +55,17 @@ install_eplus <- function (ver = "latest", force = FALSE) {
     message(sprintf("Starting to download EnergyPlus v%s...", ver), "\n", cli::rule(line = 2))
     dl <- download_eplus(ver = ver, dir = tempdir())
     exec <- attr(dl, "file")
+    data.table::setattr(exec, "version", ver)
 
     message(sprintf("Starting to install EnergyPlus v%s...", ver), "\n", cli::rule(line = 2))
+
+    message("NOTE: Administrative privileges required during installation. ",
+            "Please make sure R is running with an administrator acount or equivalent.")
+
     res <- switch(os_type(),
            windows = install_eplus_win(exec),
            linux = install_eplus_linux(exec),
            macos = install_eplus_macos(exec))
-
-    message("NOTE: Administrative privileges required during the installation. ",
-            "Please make sure R is running with an administrator acount or equivalent.")
 
     if (res != 0L) {
         stop(sprintf("Failed to install EnergyPlus v%s.", ver), call. = FALSE)
@@ -262,16 +267,22 @@ eplus_latest_release <- function () {
 # }}}
 # install_eplus_win {{{
 install_eplus_win <- function (exec) {
+    # change working directory
+    ori_wd <- getwd()
+    on.exit(setwd(ori_wd), add = TRUE)
+
+    exe_dir <- dirname(exec)
+    setwd(exe_dir)
+
     has_ps <- unname(Sys.which("powershell") != "")
     win_exec <- normalizePath(exec)
 
     if (has_ps) {
         cmd <- sprintf("& %s /S | Out-Null", exec)
-        res <- run_cmd(command = "powershell", args = cmd,
-                       echo = TRUE, windows_verbatim_args = TRUE)
+        res <- system2(command = "powershell", args = cmd)
     } else {
         message("Cannot install EnergyPlus silently. Will try to install from GUI.")
-        res <- run_cmd(commandline = exec)
+        res <- system(command = exec)
     }
 
     return(res)
@@ -279,27 +290,32 @@ install_eplus_win <- function (exec) {
 # }}}
 # install_eplus_macos {{{
 install_eplus_macos <- function (exec) {
-    no_ext <- tools::file_path_sans_ext(base)
-    cmd <- sprintf("sudo hdiutil attach %s | sudo installer -pkg /Volumes/%s/%s.pkg -target /", exec, no_ext, no_ext)
-    run_cmd(commandline = cmd)
+    # change working directory
+    ori_wd <- getwd()
+    on.exit(setwd(ori_wd), add = TRUE)
+
+    exe_dir <- dirname(exec)
+    setwd(exe_dir)
+
+    f <- basename(exec)
+    no_ext <- tools::file_path_sans_ext(f)
+    cmd <- sprintf("sudo hdiutil attach %s | sudo installer -pkg /Volumes/%s/%s.pkg -target LocalSystem",
+        f, no_ext, no_ext)
+    system(cmd)
 }
 # }}}
 # install_eplus_linux {{{
 install_eplus_linux <- function (exec) {
-    cmd <- sprintf("sudo chmod +x %s | echo 'y\r' | sudo ./%s", exec, exec)
-    run_cmd(commandline = cmd)
-}
-# }}}
-# run_cmd {{{
-# only return the status and suppress annoying warnings from `processx::run`
-run_cmd <- function (...) {
-    res <- tryCatch(
-        {r <- suppressWarnings(processx::run(...))
-         r$status
-        },
-        error = function (e) 1L
-    )
+    # change working directory
+    ori_wd <- getwd()
+    on.exit(setwd(ori_wd), add = TRUE)
 
-    return(res)
+    exe_dir <- dirname(exec)
+    setwd(exe_dir)
+
+    f <- basename(exec)
+    v <- gsub("\\.", "-", attr(exec, "version"))
+    system(sprintf("chmod +x %s", f))
+    system(sprintf('echo "y\r" | sudo ./%s | sudo chmod -R a+w /usr/local/EnergyPlus-%s', f, v))
 }
 # }}}