+++ NEWS.md
@@ -1,4 +1,4 @@
-# eplusr 0.8.2.1
+# eplusr 0.8.3
 
 ## New features